Fix several things
