+++ kospeech/dataclasses/__init__.py
@@ -1,0 +1,25 @@
+# -*- coding: utf-8 -*-
+# Soohwan Kim, Seyoung Bae, Cheolhwang Won.
+# @ArXiv : KoSpeech: Open-Source Toolkit for End-to-End Korean Speech Recognition
+# This source code is licensed under the Apache 2.0 License license found in the
+# LICENSE file in the root directory of this source tree.
+
+from .evaluate import EvalConfig
+from .audio import (
+    MfccConfig,
+    MelSpectrogramConfig,
+    SpectrogramConfig,
+    FilterBankConfig,
+)
+from .model import (
+    DeepSpeech2Config,
+    ListenAttendSpellConfig,
+    TransformerConfig,
+    JointCTCAttentionLASConfig,
+    JointCTCAttentionTransformerConfig,
+)
+from .train import (
+    DeepSpeech2TrainConfig,
+    ListenAttendSpellTrainConfig,
+    TransformerTrainConfig,
+)
